Refactored code and moved files around.
